eyeD3: Use a 'Year' tag that common applications will read.
eyeD3 currently uses TDRL 'release year' for CDDB's Date
entry which applications such as vlc and Audacious ignore.
This patch allows the same value to also be written as
TDRC 'recording date' which vlc and Audacious will read.

Thanks to Reuben Thomas for finding the issue and
demonstrating the fix. This closes Issue 126.