Fix for broken usepipes
'Usepipes' has been broken for many years, this patch simply
represents a partial revert for the original breakage:

Commit: 900a8efeddb238a1e67fec41857e3d2594227dde

This allows usepipes to run again but exposes abcde to the
issue on macosx10 which this commit tried to rectify:

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=527191

So some more attention will be required here. Tested with
cdparanoia and cdda2wav which were both fine, I presume
icedax is also fine. dagrab failed but I see that is also
broken in release version 2.6. Pird and cddafs remain
to be tested.