Album art: screen for 1x1 gif from Amazon
When there is no cover image on Amazon there should be an error
message. Instead a 1x1 pixel gif is provided for download which
is used by abcde as album art. This patch provides a simple size
test which will delete the troublesome gif and then allow glyrc
to download. Thanks to Johannes Gernemann for the patch which
resolves Issue 141.