fix release year code
The fix for #30 implemented in 1f1d33c64206841195a196595ed48f78952c07e6
relies on the release-event-list node from MB containing a count
attribute. However in my testing, that isn't the case, and so the
release year is never populated. This change checks that the
release_event_list isn't undef before trying to get the events (the
cause of the exception in #30) but then doesn't rely on the count() of
the release_event_list.