Fix compatibility with WebService::MusicBrainz 1.x
Closes Issue #60: https://abcde.einval.com/bugzilla/show_bug.cgi?id=60

Account for changes in WebService::MusicBrainz v1.x which uses the v2
MusicBrainz API.

This converts the API call to the v2 API, and account for the slightly
different data structure of the response.

Track offsets are directly available in the response and don't need
computation anymore.

There is some complication to locate the track names since the v2 API
response may return multiple media per release, including media with an
ID that is different from the one originally requested.